+++ samplesheets/tasks_taskflow.py
@@ -1,15 +1,24 @@
 """Taskflow tasks for the samplesheets app"""
 
+import logging
+
+from typing import Optional
+
+from django.conf import settings
+
 # Projectroles dependency
+from projectroles.models import SODARUser
 from projectroles.plugins import PluginAPI
 
 # Samplesheets dependency
 from samplesheets.models import Investigation
+from samplesheets.tasks_celery import update_project_cache_task
 
 # Taskflowbackend dependency
 from taskflowbackend.tasks.sodar_tasks import SODARBaseTask
 
 
+logger = logging.getLogger(__name__)
 plugin_api = PluginAPI()
 
 
@@ -54,3 +63,29 @@ class RemoveSampleSheetsTask(SODARBaseTask):
 
     def revert(self, *args, **kwargs):
         pass  # TODO: How to handle this?
+
+
+class UpdateProjectSheetCacheTask(SODARBaseTask):
+    """Update project sample sheet cache"""
+
+    def execute(
+        self,
+        user: Optional[SODARUser],
+        add_alert: bool,
+        alert_msg: Optional[str],
+        *args,
+        **kwargs,
+    ):
+        if settings.SHEETS_ENABLE_CACHE:
+            try:
+                update_project_cache_task.delay(
+                    project_uuid=str(self.project.sodar_uuid),
+                    user_uuid=str(user.sodar_uuid) if user else None,
+                    add_alert=add_alert,
+                    alert_msg=alert_msg,
+                )
+            except Exception as ex:
+                logger.error(f'Unable to run project cache update task: {ex}')
+        super().execute(*args, **kwargs)
+
+    # NOTE: No revert needed